Store bash history in tooling container
Since xterm.js does not save history on disconnecting, it's necessary to
configure the tooling container to write history after every command. To
do this, we set PROMPT_COMMAND to sync the current session with the
.bash_history file every time the prompt is printed.

This preserves history for the runtime of a web terminal pod. History is
still lost whenever the pod is deleted or idled.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>