Add kn wrapper script to allow downloading kn from the cluster
Add a wrapper script for the 'kn' CLI, which can be used to download
the kn CLI from the cluster instead of using the built-in kn binary.

This is necessary as using a version of kn that doesn't match the
version of the serverless operator installed in a cluster causes issues.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>