Add wrapper for tkn similar to kn-wrapper
When a user runs the tkn CLI in a cluster with OpenShift Pipelines
installed, prompt the user to optionally download the tkn CLI (and
tkn-pac, opc) from the operator in the cluster to ensure the CLI version
matches the operator's version.

The user's preference is saved between runs of the terminal via a
DevWorkspace annotation.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>