Fix issue in detecting if logged in to a cluster
Executing 'oc whoami -c' fails if no actual user is logged in, even if
there is a serviceaccount token mounted to the container. In this case,
the serviceaccount can still be used to make API calls even though
there's no context set.

This causes automatic restoring of wrapper preferences to fail, as at
the time the entrypoint is executed, no actual user is logged in.

Instead, check if there is any OpenShift token available (oc whoami -t)

Signed-off-by: Angel Misevski <amisevsk@redhat.com>