Update to ubi-minimal:9.4-1134
oc 4.16 requires GLIBC_2.33 & GLIBC_2.34 which are only available on RHEL 9 based images.
Due to this dependency requirement, we are updating from UBI8 to UBI9.

The curl-minimal package is also provided by default in the UBI9 image, and installing
curl manually results in package conflicts. Thus we are no longer manually installing
curl in the tooling container.

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>